# openapi.yaml
openapi: 3.0.0
info:
  title: School Timetable Management API with Wellness
  version: 2.0.0
  description: |
    RESTful API for the School Timetable Management SaaS Platform, including automated timetable generation, 
    teacher wellness monitoring, workload management, and substitution handling. Supports multi-tenant schools 
    with role-based access (admin, principal, teacher, student, parent).
servers:
  - url: https://api.schooltimetable.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Teacher:
      type: object
      required: [userId, subjects]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        subjects:
          type: array
          items:
            type: string
        availability:
          type: object
        preferences:
          type: object
        maxPeriodsPerDay:
          type: integer
          default: 6
        maxPeriodsPerWeek:
          type: integer
          default: 30
        maxConsecutivePeriods:
          type: integer
          default: 3
        minBreakDuration:
          type: integer
          default: 10
        wellnessScore:
          type: number
        burnoutRiskLevel:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
...
    Class:
      type: object
      required: [name, grade]
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        name:
          type: string
        grade:
          type: integer
        section:
          type: string
        stream:
          type: string
          enum: [Science, Commerce, Arts]
        studentCount:
          type: integer
    Subject:
      type: object
      required: [name, credits]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        department:
          type: string
        credits:
          type: integer
        minPeriodsPerWeek:
          type: integer
        maxPeriodsPerWeek:
          type: integer
        prepTime:
          type: integer
        correctionWorkload:
          type: number
        requiresLab:
          type: boolean
...
    Room:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        capacity:
          type: integer
        type:
          type: string
          enum: [classroom, lab, auditorium, office]
...
    Timetable:
      type: object
      required: [schoolId, academicYearId]
      properties:
        id:
          type: string
          format: uuid
        schoolId:
          type: string
          format: uuid
        academicYearId:
          type: string
          format: uuid
        version:
          type: integer
        status:
          type: string
          enum: [draft, approved, active, archived]
        wellnessScore:
          type: number
        workloadBalanceScore:
          type: number
        metadata:
          type: object
        wellnessAnalysis:
          type: object
        createdBy:
          type: string
          format: uuid
        approvedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    TimetableWithWellness:
      type: object
      properties:
        timetable:
          $ref: '#/components/schemas/Timetable'
        wellnessAnalysis:
          type: object
          properties:
            overallScore:
              type: number
            teacherScores:
              type: object
            departmentScores:
              type: object
            issues:
              type: array
              items:
                type: object
            improvements:
              type: array
              items:
                type: string
...
paths:
...
  /teachers:
    get:
      summary: List teachers
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of teachers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Teacher'
    post:
      summary: Create teacher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Teacher'
      responses:
        201:
          description: Teacher created
...
  /classes:
    get:
      summary: List classes
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Class'
    post:
      summary: Create class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Class'
      responses:
        201:
          description: Class created
...
  /subjects:
    get:
      summary: List subjects
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subject'
    post:
      summary: Create subject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subject'
      responses:
        201:
          description: Subject created
...
  /rooms:
    get:
      summary: List rooms
      parameters:
        - name: schoolId
          in: query
          schema:
            type: string
          format: uuid
      responses:
        200:
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
    post:
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Room'
      responses:
        201:
          description: Room created
...
  /timetables/{id}/generate:
    post:
      summary: Generate timetable with wellness optimization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                options:
                  type: integer
                  default: 3
                timeout:
                  type: integer
                  default: 60
                optimization:
                  type: object
                  properties:
                    academicWeight:
                      type: number
                      default: 0.35
                    wellnessWeight:
                      type: number
                      default: 0.30
                    efficiencyWeight:
                      type: number
                      default: 0.20
                    preferenceWeight:
                      type: number
                      default: 0.15
      responses:
        200:
          description: Generated timetable options with wellness scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimetableWithWellness'
...